Fix type casting from uint8 to int8
int8_t has half the values of uint8_t as it needs to account for the
sign, thus causing the binary threshold cache to overflow and go
negative, therefore also causing the computation to happen on every
single new pixel read. This commit just converts the cached value type
to int16_t so that there's no overflow risk at the cost of an extra
byte. It also fixes an incorrect default setting on bmps with <
24bpp.