Allow fetching tfl line status in a single call
The old implementation required individual api calls for each status.
This new method allows chaining of lines in a single API request. This
has the downside of making the response larger, but still fits
comfortabily in memory and has a massive speed up (not to mention that
it lowers the power usage considerably).