Add basic input management
This adds support for a rotary encoder hooked up to CLK(5), DT(4) and
SW(33). The GPIO 33 is chosen so that it can be used as a RTC input to
wake up from deep sleep.

For now this doesn't do anything particularly useful, but we should
detect clicks & long clicks to refresh and enter configuration mode.